% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelnice-loading.R
\name{load_strelnice_results}
\alias{load_strelnice_results}
\title{Loads data from passed path}
\usage{
load_strelnice_results(filepth)
}
\arguments{
\item{pth}{}
}
\value{

}
\description{
Loads data from passed path
}
